==> workflow/04c-ARGs-interactions-relative.Rmd <==
---
title: "04c-ARGs-interactions-relative"
date: "Compiled at `r format(Sys.time(), '%Y-%m-%d %H:%M:%S', tz = 'UTC')` UTC"
output: github_document
params:
  name: "04c-ARGs-interactions-relative" # change if you rename file
---

```{r here, message=FALSE, echo = F}
here::i_am(paste0(params$name, ".Rmd"), uuid = "56797c4b-eb48-4d32-bf93-a767bb32d50b")
knitr::opts_chunk$set(dpi = 200, echo = T, warning = F, message = F)
```


Here we treat the counts as relative abundances.

```{r packages}
library("conflicted")
library(dplyr)
library(tidyr)
library(textshape)
library(ggplot2)
#devtools::install_github("viettr/trac", dependencies = TRUE, ref = "prediction_during_cv")
library(trac)
library(gridExtra)
```

```{r directories, echo = F}
# create or *empty* the target directory, used to write this file's data: 
projthis::proj_create_dir_target(params$name, clean = TRUE)

# function to get path to target directory: path_target("sample.csv")
path_target <- projthis::proj_path_target(params$name)

# function to get path to previous data: path_source("00-import", "sample.csv")
path_source <- projthis::proj_path_source(params$name)
```

### Read data

```{r}
path_data <- "data/"
mOTU_all <- readRDS(paste0(path_data, "mOTU_all.rds"))
meta_all <- readRDS(paste0(path_data, "Metadata_all.rds"))
```


```{r}
## extract genus level and adjust names
mOTU_genus <- mOTU_all$Genus
rownames(mOTU_genus) <- substr(rownames(mOTU_genus), 4, nchar(rownames(mOTU_genus)))

mOTU_genus[1:5, 1:5]
```


1. Remove samples with missing data in the metadata


```{r}
dim(meta_all)
```

Read ARG data and add it to meta data

```{r}
arg_df <- read.table(paste0(path_data, "hub.microbiome.summary.down.10000000.r"), sep='\t')
arg_df <- arg_df %>% 
  pivot_wider(id_cols = "SampleID", names_from="Feature", values_from = "FeatureValue") %>% 
  select(SampleID, CARD10M) # For now only interested in ARGs

head(arg_df)
```

```{r}
meta_arg <- meta_all %>% 
  tibble::rownames_to_column("SampleID") %>% 
  left_join(arg_df, by="SampleID") %>% 
  tibble::column_to_rownames("SampleID")
meta_all <- meta_arg
```


```{r}
meta_all.f = meta_all[complete.cases(meta_all),]
dim(meta_all.f)
```


```{r}
ind_genus = intersect(rownames(meta_all.f), rownames(mOTU_genus))
length(ind_genus)
dim(mOTU_genus)
```

```{r}
## only merged / intersection
meta_all.f.m <- meta_all.f[ind_genus, ]
mOTU_genus.m <- mOTU_genus[ind_genus, ]

dim(meta_all.f.m)
dim(mOTU_genus.m)
```


2. Remove covariates with only zeros


```{r}
sum(colSums(meta_all.f.m!= 0) == 0)
meta_all.f.m = meta_all.f.m[, colSums(meta_all.f.m!= 0) > 0]
```



Let's take into account the 30 most abundant genera, remove "unclassified":

```{r}
order_abund <- order(colSums(mOTU_genus.m), decreasing = T)
X <- mOTU_genus.m[, order_abund[2:33]]
## remove duplicates
X <- X[, !(colnames(X) %in% c("_Ruminococcus_", "_Bacteroides_"))]
colnames(X)
```

As y we choose the number of ARGs

```{r}
y <- meta_all.f.m$CARD10M
names(y) <- rownames(meta_all.f.m)
all(names(y) == rownames(X))
```

Make sure there are no special characters in the names

```{r}
colnames(X) <- gsub("_", "", colnames(X))
```


```{r}

pseudo_count <- 1

Xm <- as.matrix(X + 0.00001 * matrix(rnorm(length(X)), nrow = nrow(X), ncol = ncol(X)))
if(sum(Xm == 0) > 0){X_psd <- as.matrix(Xm + pseudo_count)}
X_psd <- as.matrix(Xm + pseudo_count)
#X_rel <- Xm/rowSums(Xm)
X_rel <- X_psd/rowSums(X_psd)
```

### Fit SLC and SLC + int model

```{r message=F, warning=F}
source("R/sparse_log_contrast.R")
source("R/slc_int.R")
source("R/slc_int_plots.R")
source("R/log-ratio-lasso_interactions.R")
source("R/utils.R")
```

```{r}
library(trac)
```


```{r eval = T, results='hide'}
slc_slc_int_ARGs <- slc_slc_int_all_splits(X = as.matrix(X_rel),
                                             y = y,
                                             method = "regr", output = "raw", ii = "ibest")
```

### OOS MS





```{r}
nsplit = 10 
rsq <- function (x, y) cor(x, y) ^ 2

list_plt_slcint <- list()
tr <- list()
yhat_te <- list()

for (r in seq(nsplit)) {
  yhat_te[[r]] <- slc_slc_int_ARGs$yhat_te_int[[r]][, slc_slc_int_ARGs$cvfit[[r]]$cv$ibest]
  tr[[r]] <- slc_slc_int_ARGs$tr[[r]]
  data <- data.frame(Observed = y[-tr[[r]]], Predicted = as.vector(yhat_te[[r]]))
  
  list_plt_slcint[[r]] <- ggplot(data, aes(x = Observed, y = Predicted)) +
    geom_point(alpha = 0.5) +  # Set alpha for transparency
    geom_abline(intercept = 0, slope = 1, color = "red") +
    labs(title = paste0("slcInt: Train-test split ", r), y = "Predicted (test)", x = "Observed (test)") +
    annotate("text", x = max(data$Observed), y = min(data$Predicted),
             label = paste0("R-squared: ", round(rsq(data$Observed, data$Predicted), 2)),
             vjust = 1, hjust = 1) +
    theme_minimal()
}



# Arrange the ggplots in a single row
# grid.arrange(grobs = list_plt_slcint[1:3], ncol = 3)
```

```{r echo = F,eval = F}
#saveRDS(t(slc_slc_int_ARGs$beta_int_est_refit), "coef_qlc_relative_ARGs.rds")
```


```{r echo = F, eval = T}
# Build superset of all nonzero coefficients in all three models
s1 <- readRDS("data/selected-features-APL-counts.rds")
s2 <- readRDS("data/selected-features-APL-binary.rds")
s3 <- readRDS("data/selected-features-qlc-compositions.rds")
sel12 <- union(s1, s2)
sel123 <- union(sel12, s3)
non_zero_columns <- t(slc_slc_int_ARGs$beta_int_est_refit)[, sel123]

# Identify columns with and without ":"
selected_columns <- grep(":", colnames(non_zero_columns), value = TRUE)
# Apply scaling factor to interaction columns
non_zero_columns[, selected_columns] <- non_zero_columns[, selected_columns] * 20

# Melt the data for ggplot
data_long <- reshape2::melt(as.data.frame(non_zero_columns))
order_by_median <- readRDS("data/universal_order_boxplots.rds")
data_long$variable <- factor(data_long$variable, levels = rev(order_by_median))

# Assign colors based on the presence of ":"
# Prepare color assignment dynamically
data_long$color <- ifelse(grepl(":", data_long$variable), "darkred", "lightblue")

# Plotting
plt_coef <- ggplot(data_long, aes(x = value, y = variable, fill = color)) +
  geom_boxplot() +
  labs(x = "Estimated coefficients 10 splits", y = "Features", title = "Boxplot of Estimated Coefficients (10 splits)") +
  theme_minimal() +
  theme(axis.text.y = element_text(angle = 0, vjust = 0.5, hjust = 1, size = 12, color = "black"),
        axis.text.x = element_text(size = 12, color = "black"),
        legend.position = "none") +
  scale_fill_manual(values = c("lightblue" = "lightblue", "darkred" = "darkred"))  # Ensure colors are specified here

# Optionally save or display the plot
# ggsave(paste0(path.temp, "boxplot-coef-MetaCardis-ARGs-qlc-comp.pdf"), plt_coef, height = 10, width = 7)
```



```{r, echo=F, fig.width=15, fig.height=20}
plt_coef
```





## Files written

These files have been written to the target directory, ```r paste0("data/", params$name)```:

```{r list-files-target}
projthis::proj_dir_info(path_target())
```
